- Add explicit template instantiations for:     - vsg::Array<vsg::DrawIndexedIndirectCommand>     - vsg::Array<vsg::VkGeometryInstance> - cast l.size() to uint32_t in assign() to match parameter type and avoid warnings - rename core/Array.cpp to core/ArrayInstantiations.cpp

+++ include/vsg/core/Array.h
@@ -102,7 +102,7 @@ namespace vsg
             _data(nullptr),
             _size(0)
         {
-            assign(data, offset, stride, l.size());
+            assign(data, offset, stride, static_cast<uint32_t>(l.size()));
 
             iterator itr = begin();
             for (const value_type& v : l) { (*itr++) = v; }
